Refactor `Lexer` and `SourceFile` buffer handling for improved consistency
- Replace `end_` with `endBuffer_` in `Lexer` for clearer pointer semantics.
- Add `TRAILING_0` constant in `SourceFile` to define the forced byte count.
- Introduce `SourceFile::size()` for buffer size calculation excluding `TRAILING_0`.
- Use a loop for trailing zero initialization in `SourceFile::loadContent`.
- Update `swc.cpp` to cast buffer size to `uint32_t` for compatibility.

+++ src/Main/swc.cpp
@@ -28,7 +28,7 @@ static void parseFolder(CompilerContext& ctx, const fs::path& directory)
             {
                 const auto f = new SourceFile(entry.path());
                 f->loadContent(ctx);
-                if (f->codeView(0, f->content().size()).find("#global testerror") == Utf8::npos)
+                if (f->codeView(0, (uint32_t) f->content().size()).find("#global testerror") == Utf8::npos)
                 {
                     ctx.setSourceFile(f);
                     f->tokenize(ctx);
